+++ lib/karafka/setup/config.rb
@@ -44,9 +44,10 @@ module Karafka
       setting :batch_processing, false
       # Should we operate in a single controller instance across multiple batches of messages,
       # from the same partition or should we build a new instance for each incoming batch.
-      # Persistent mode can be usefull when we want to flush data after multiple messages were
-      # received and processed together
-      setting :persistent, false
+      # Disabling that can be useful when you want to build a new controller instance for each
+      # incoming batch. It's disabled by default, not to create more objects that needed on
+      # each batch
+      setting :persistent, true
       # Connection pool options are used for producer (Waterdrop)
       # They are configured automatically based on Sidekiq concurrency and number of consumers
       # The bigger one is selected as we need to be able to send messages from both places