+++ spec/lib/karafka/routing/router_spec.rb
@@ -3,46 +3,31 @@
 RSpec.describe Karafka::Routing::Router do
   subject(:router) { described_class }
 
-  let(:partition) { rand(100) }
-  let(:kafka_message) do
-    instance_double(
-      Kafka::FetchedMessage,
-      partition: partition,
-      topic: topic_name
-    )
-  end
+  before do
+    Karafka::Routing::Builder.instance.delete_if { true }
 
-  describe 'build' do
-    context 'not persisted' do
-      context 'when we want to build a controller out of non existing topic' do
-        let(:group_id) { rand.to_s }
-        let(:topic_name) { rand.to_s }
-        let(:expected_error) { Karafka::Errors::NonMatchingRouteError }
-
-        it { expect { router.build(group_id, kafka_message) }.to raise_error(expected_error) }
-      end
-
-      context 'when we want to build controller out of an existing topic' do
-        let(:group_id) { Karafka::Routing::Builder.instance[0].topics[0].consumer_group.id }
-        let(:topic_name) { 'topic_name1' }
-
-        before do
-          Karafka::Routing::Builder.instance.draw do
-            topic :topic_name1 do
-              controller Class.new(Karafka::BaseController)
-              processing_backend :inline
-            end
-          end
-        end
-
-        it do
-          expect(router.build(group_id, kafka_message)).to be_a(Karafka::BaseController)
-        end
+    Karafka::Routing::Builder.instance.draw do
+      topic :topic_name1 do
+        controller Class.new(Karafka::BaseController)
+        persistent false
+        batch_processing true
       end
     end
+  end
 
-    context 'persisted' do
-      pending
+  describe 'find' do
+    context 'non existing topic' do
+      let(:topic_id) { rand.to_s }
+      let(:expected_error) { Karafka::Errors::NonMatchingRouteError }
+
+      it { expect { router.find(topic_id) }.to raise_error(expected_error) }
+    end
+
+    context 'existing topic' do
+      let(:topic) { Karafka::Routing::Builder.instance.last.topics.last }
+      let(:topic_id) { topic.id }
+
+      it { expect(router.find(topic_id)).to eq topic }
     end
   end
 end